+++ src/components/Modal/styles.ts
@@ -1,0 +1,63 @@
+import styled from "styled-components";
+
+////Modal
+export const ModalContainer = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: fixed;
+  `;
+
+export const DialogBox = styled.dialog`
+  width: 662px;
+  height: 497px;
+  margin-top: -200px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  border: none;
+  border-radius: 14px;
+  box-shadow: 0 0 30px rgba(30, 30, 30, 0.185);
+  box-sizing: border-box;
+  background-color: white;
+  z-index: 10000;
+  padding: 62px;
+`;
+
+export const Backdrop = styled.div`
+  width: 100vw;
+  height: 100vh;
+  position: fixed;
+  top: 0;
+  z-index: 9999;
+  background-color: rgba(0, 0, 0, 0.2);
+`;
+
+export const BackdropBtn = styled.button`
+  width: 30px;
+  height: 30px;
+  font-size: 30px;
+  text-align: center;
+  line-height: 30px;
+  background-color: white;
+  border: none;
+  color: #d9d9d9;
+  margin: 0;
+  position: absolute;
+  right: 32px;
+  top: 32px;
+  cursor: pointer;
+  &:hover {
+    color: black;
+  }
+`;
+
+export const Children = styled.div`
+  width: 538px;
+  height: 372px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
